+++ src/nodes/statements/generic_for.rs
@@ -100,6 +100,11 @@ impl GenericForStatement {
     }
 
     #[inline]
+    pub fn mutate_tokens(&mut self) -> Option<&mut GenericForTokens> {
+        self.tokens.as_mut()
+    }
+
+    #[inline]
     pub fn get_block(&self) -> &Block {
         &self.block
     }