+++ src/nodes/statements/repeat_statement.rs
@@ -79,6 +79,11 @@ impl RepeatStatement {
         self.tokens.as_ref()
     }
 
+    #[inline]
+    pub fn mutate_tokens(&mut self) -> Option<&mut RepeatTokens> {
+        self.tokens.as_mut()
+    }
+
     pub fn clear_comments(&mut self) {
         if let Some(tokens) = &mut self.tokens {
             tokens.clear_comments();